Improve sidebar visibility and abbreviate month in header
- Increase active tab background opacity (25% dark, 12% light)
- Make active tab left border thicker (3px → 4px)
- Add visible green hover background (15% dark, 8% light)
- Add green border on hover (50% opacity)
- Change text and icon color to green on hover
- Abbreviate month in dashboard header to 3 letters (short format)

+++ farmtrackr-web/src/app/DashboardClient.tsx
@@ -498,7 +498,7 @@ export default function DashboardClient({ contacts, stats }: DashboardClientProp
                     margin: '0 0 2px 0',
                     lineHeight: '32px'
                   }}>
-                    {currentDateTime.toLocaleDateString('en-US', { month: 'long', day: 'numeric' })}
+                    {currentDateTime.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })}
                   </p>
                   <p style={{ 
                     fontSize: '11px',
